+++ .eslintignore
@@ -1,1 +1,2 @@
 index.html
+/example/src/assets/js/SHA1.js
