+++ lib/check.js
@@ -6,7 +6,7 @@ let _appId,// 唯一标识
   _nonceStr,// 生成签名的随机串
   _signature,// 签名
   _errorHandler
-  // _event
+// _event
 
 /**
  * 判断一个字符串是否为空，如果非字符串参数，则抛出类型错误异常
@@ -16,9 +16,9 @@ let _appId,// 唯一标识
 function utilStrIsEmpty(str) {
   if (str && !is('String', str)) {
     if(is('Function', _errorHandler)) {
-      _errorHandler(`请确认给定参数值: ${str} 是否为字符串`)
+      _errorHandler(`请确认给定参数: ${str} 是否为字符串`)
     }
-    throw new Error(`请确认给定参数值: ${str} 是否为字符串`)
+    throw new Error(`请确认给定参数: ${str} 是否为字符串`)
   }
   return (!str || str.trim().length === 0)
 }
