+++ lib/index.js
@@ -49,5 +49,4 @@ export class JSBridge {
       })
     })
   }
-
 }
