+++ lib/js-bridge-context.js
@@ -1,4 +1,3 @@
-/* eslint-disable */
 import {is} from './util/type'
 import device from './util/device'
 import {emitErr} from './util/warn'
@@ -19,7 +18,6 @@ let canUseJsbridgeInterface = true
 function _handlerResp(res, resolve, reject) {
   try {
     const temp = window.JSON.parse(res)
-    console.log(temp)
     if (temp.ReturnCode === '000000') {
       resolve(temp)
     } else {
