--- test1.html
@@ -1,53 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>123</title>
-    <!--<script src="/dist/build.js"></script>-->
-    <script src="./ynrcc-mobilebank-jssdk.js"></script>
-   <script type="text/javascript" src="./SHA1.js"></script>
-    <script type="text/javascript" src="./comm.js"></script>
-    <script type="text/javascript">
-		  function back() {
-	        jsBridge.closeWindow();
-	      }
-      function getUserInfo() {
-        jsBridge.getUserInfo().then(res => {
-          alert(JSON.stringify(res))
-        }).catch(err => {
-          alert(JSON.stringify(err))
-        })
-      }
-      function test() {
-        document.activeElement.blur();
-        jsBridge.showPwdKeyboard(document.getElementById("pwd"), "password");
-      }
-      function getPwd() {
-        jsBridge.getEncrypt(['password'], new Date().getTime()).then(res => {
-          alert(JSON.stringify(res))
-        }).catch(error => {
-          alert('showPwdKeyboard err ' + error.message)
-        })
-      }
-      function clearPwd() {
-        jsBridge.onTapClearKeyboardEncrypt(document.getElementById("pwd"), "password")
-      }
-      function getMobileDevice() {
-        jsBridge.getMobileDevice().then(res => {
-          alert(JSON.stringify(res))
-        })
-      }
-    </script>
-
-</head>
-
-<body>
-  <a onclick="getMobileDevice()">返回手机银行</a><br><br>
-  <a onclick="share()">微信分享</a><br><br>
-  <a onclick="getUserInfo()">获取用户信息</a><br><br>
-  <a onclick="getPwd()">获取密文</a><br><br>
-  <a onclick="clearPwd()">清除密文</a><br><br>
-  <br/><br/><br/><br/>
-  <input id="pwd" onfocus="test()">
-</body>
-</html>
